Use local path for file args to `java` for tramp support
`lsp-java-server-install-dir` (for example) is user config that can be
set to remote directory in user config. But, the command takes the
remote path as is and passes to java command which fails to resolve the
path. Using local file for the command running on remote server solves
the problem and starts server normally.