Ignore missing JAR in native installations
Do not check for the existence of a JAR if a native command is preferred
and there is a suitable executable.

I was hit by this issue on NixOS which puts the JAR at a different
location than the auto detection of this package expects it. The
alternative, setting `lsp-java-server-install-dir` such that the JAR
instead of the executable is found, leads to a failure. When the JAR is
launched directly, jdtls tries to write to the Nix Store which is
read-only. Launching the jdtls executable provided by nixpkgs works
perfectly though.